VideoMAEv2/requirements.txt
packaging
ninja
av==11.0.0
decord==0.6.0
deepspeed==0.10.1
einops==0.7.0
flash_attn
fvcore==0.1.5.post20221221
imageio==2.31.1
librosa==0.10.1
numpy==1.24.4
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
PyYAML==6.0.1
scipy
soundfile==0.12.1
termcolor==2.4.0
timm==0.5.4
#torch==1.13.1+cu118
#torchaudio==0.13.1+cu118
#torchvision==0.14.1+cu118
tqdm==4.66.1
transformers==4.28.1
wandb==0.16.1
peft
open-clip-torch
huggingface_hub